// frontend/src/api/generated/model/farmViewModel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { GeoCoordinateViewModel } from './geoCoordinateViewModel';

export interface FarmViewModel {
  id: string;
  name: string;
  city: string;
  state: string;
  headquarters: null | GeoCoordinateViewModel;
  createdAt: string;
  /** @nullable */
  updatedAt: string | null;
}